abstract camera update logic to ECS system
