+++ src/systems/mod.rs
@@ -1,2 +1,4 @@
 pub mod camera;
 pub mod prepare;
+
+pub use camera::camera_system;
